refactor: move media helper function into its own file
